+++ shared/src/commonMain/kotlin/dev/dhun/ui/browse/PlaylistScreen.kt
@@ -1,6 +1,7 @@
 package dev.dhun.ui.browse
 
 import androidx.compose.foundation.background
+import androidx.compose.foundation.border
 import androidx.compose.foundation.clickable
 import androidx.compose.foundation.layout.Arrangement
 import androidx.compose.foundation.layout.Box
@@ -42,6 +43,7 @@ import dev.dhun.design.DhunIconView
 import dev.dhun.design.DhunShapes
 import dev.dhun.design.DhunSpacing
 import dev.dhun.design.components.ArtworkImage
+import dev.dhun.design.components.lyricsVeilBrush
 import dev.dhun.design.components.DhunButton
 import dev.dhun.design.components.DhunIconButton
 import dev.dhun.design.components.DhunOutlinedButton
@@ -119,7 +121,8 @@ fun PlaylistScreen(
             modifier = Modifier
                 .padding(DhunSpacing.sm)
                 .clip(DhunShapes.full)
-                .background(DhunColors.glassStrong)
+                .background(lyricsVeilBrush())
+                .border(androidx.compose.foundation.BorderStroke(DhunSpacing.border, DhunColors.glassEdge), DhunShapes.full)
                 .padding(DhunSpacing.xs),
         ) {
             DhunIconButton(